requirements.txt
pandas==0.23.4
torch==1.0.0
torchvision==0.2.1
numpy==1.15.4
Pillow==5.1.0
matplotlib==3.0.2
